Refactored views/urls. Files were getting long
