+++ backend/api/views/scenario_view.py
@@ -1,46 +1,15 @@
-import api
-import uuid
-from django.db.models import Q
-from django.http import JsonResponse
-from django.conf import settings
-from django.shortcuts import get_object_or_404
-from django.contrib.auth import authenticate, login, logout
-from django.contrib.auth.models import User
-from django.contrib.auth.decorators import login_required
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from .serializers import PatientSerializer, ScenarioSerializer, MedicationSerializer
-from .models import Patient, Scenario, Medication, Scenario_Medication
 from datetime import datetime
 from django.db import transaction
-
-
-# Create your views here.
-
-@api_view(['GET'])
-def application_data(request):
-    data = {
-        "appName": "Test Data",
-        "version": "1.0.0",
-        "description": "This is CORS Test Data."
-    }
-    return JsonResponse(data)
-
-@api_view(['POST'])
-def submit_patient(request):
-    print(request.data)
-    serializer = PatientSerializer(data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-        return Response({'message': 'Form submitted successfully'}, status=200)
-    return Response(serializer.errors, status=400)
-
-@api_view(['GET'])
-def get_patient(request, patient_id):
-    print(request)
-    patient = get_object_or_404(Patient, patient_id=patient_id)
-    serializer = PatientSerializer(patient)
-    return Response(serializer.data, status=200)
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+from ..serializers import PatientSerializer, ScenarioSerializer
+from ..models import Patient, Medication, Scenario_Medication
+from .decorators import login_required
+from django.db.models import Q
+from ..models import Scenario
+from django.http import JsonResponse
+from django.shortcuts import get_object_or_404
+import uuid
 
 # Helper function to generate a unique scenario ID
 def generate_unique_scenario_id():
@@ -51,6 +20,7 @@ def generate_unique_scenario_id():
         # Check if the ID already exists in the database
         if not Scenario.objects.filter(scenario_id=scenario_id).exists():
             return scenario_id
+
 
 @api_view(['POST'])
 @login_required
@@ -112,9 +82,8 @@ def submit_scenario(request):
             except Exception as e:
                 print(f"Error creating medication: {e}")
                 return Response({'error': f"Failed to process medication: {med_data}"}, status=400)
-
         scenario_data = {
-            "scenario_id": generate_unique_scenario_id(),
+            "scenario_id":generate_unique_scenario_id(),
             "owner": request.user.id,
             "name": data.get("name"),
             "description": data.get("description"),
@@ -140,49 +109,8 @@ def submit_scenario(request):
 
     return Response({'message': 'Form submitted successfully'}, status=200)
 
-@api_view(['POST'])
-def login_page(request):
-    # Check if the HTTP request method is POST (form submission)
-    username = request.POST.get('username')
-    password = request.POST.get('password')
-    
-    if not username or not password:
-        return JsonResponse({'success': False, 'error': 'Username and password are required'}, status=400)
-
-    username = username.lower()
-
-    # Check if a user with the provided username exists
-    if not User.objects.filter(username=username).exists():
-        return JsonResponse({'success': False, 'error': 'Invalid username'}, status=400)
-
-    # Authenticate the user with the provided username and password
-    user = authenticate(username=username, password=password)
-    
-    if user is None:
-        return JsonResponse({'success': False, 'error': 'Invalid password'}, status=400)
-    # Log in the user and redirect to the home page upon successful login
-    login(request._request, user)
-    print("Logged in")
-    return JsonResponse({'success': True, 'message': 'Login successful'})
-
-@api_view(['GET'])
-def current_user(request):
-    if request.user.is_authenticated:
-        return JsonResponse({
-            'username': request.user.username,
-            'email': request.user.email,
-        })
-    else:
-        return JsonResponse({'error': 'Not authenticated'}, status=401)
-    
-@api_view(['POST'])
-def logout_view(request):
-    auth_request = getattr(request, '_request', request)
-    logout(auth_request)
-    return JsonResponse({'success': True, 'message': 'Logged out successfully'})
-
 @login_required
-def scenario_data(request):
+def get_user_scenarios(request):
     user = request.user.id
 
     # Fetch scenarios owned by the user or the global user
@@ -195,3 +123,11 @@ def scenario_data(request):
     return JsonResponse({
         'scenarios': list(scenarios),
     })
+
+@login_required
+def get_single_scenario(request, scenario_id):
+    scenario = get_object_or_404(Scenario, scenario_id=scenario_id)
+    if request.user != scenario.owner:
+        return JsonResponse({'error': 'Scenario does not belong to this user.'}, status=401 )
+    serializer = ScenarioSerializer(scenario)
+    return JsonResponse(serializer.data)
